pyproject.toml
[project]
name = "paper-tool"
version = "0.1.0"
description = "PDF 文献自动分类与重命名工具"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0.0",
    "PyMuPDF>=1.24.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pystray>=0.19.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
paper-tool = "paper_tool.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_tool"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
